% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calendar.R
\name{g_calendar}
\alias{g_calendar}
\title{Create a R calendar}
\usage{
g_calendar(data)
}
\arguments{
\item{data}{data.frame that must contain different columns : \itemize{
  \item weekday : Day number of the current week.
  \item monthweek : Week number of the current month.
  \item valueCol : Value to be plotted in the calendar.
  \item dateCol : Date colum in format "dmy".
}}
}
\value{
\code{gg} object.
}
\description{
Create a R calendar using geom_tile from ggplot.
}
\examples{

}
\references{
all code come from https://vietle.info/post/calendarheatmap/
}
