% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_build_res_path.R
\name{get_build_res_path}
\alias{get_build_res_path}
\title{Get and build result path.}
\usage{
get_build_res_path()
}
\value{
2 actions are made
        \enumerate{
          \item gives the path name of the res folder mirror of a R script
          \item create the res path corresponding of the R script
        }
}
\description{
Executing this function in an open script in Rstudio will create a mirror repository
 in "res" folder to store results of the script.
}
\examples{

}
