% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalise_data.R
\name{normalize_data}
\alias{normalize_data}
\title{Normalize Numeric Data}
\usage{
normalize_data(x, method = "minmax")
}
\arguments{
\item{x}{A numeric vector}

\item{method}{Character, either "minmax" or "zscore". Default is "minmax"}
}
\value{
A normalized numeric vector
}
\description{
Normalizes numeric data using min-max normalization or z-score standardization.
}
\examples{
data <- c(10, 20, 30, 40, 50)
normalize_data(data, method = "minmax")
normalize_data(data, method = "zscore")
}
