% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_build_res_path.R
\name{get_build_res_path}
\alias{get_build_res_path}
\title{get and build result path}
\usage{
get_build_res_path()
}
\value{
2 actions are made
        \enumerate{
          \item gives the path name of the res folder mirror of the path of the R script
          \item create the res path corresponding of the R script
          \item graph : Barplot of detections depending of distance, with confindence
          estimated with mcmc chains.
        }
}
\description{
get and build result path
}
\examples{


}
